added Id to signup.html elements

+++ apps/recipe/views/recipe.py
@@ -1,9 +1,11 @@
-from django.shortcuts import render,  get_list_or_404, get_object_or_404, redirect
+from django.shortcuts import render, get_list_or_404
 from recipe.models import Recipe
+from django.shortcuts import get_object_or_404, redirect
 from django.contrib.auth.models import User
+from django.http import HttpResponseRedirect
 from django.contrib import auth, messages
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
-
+from django.contrib.auth.decorators import login_required
 
 
 def index(request):
@@ -14,7 +16,7 @@ def index(request):
     recipe_per_page = paginator.get_page(page)
 
     data = {
-        'recipes': recipe_per_page 
+        'recipes': recipe_per_page
     }
     return render(request, 'recipe/index.html', data)
 
@@ -41,20 +43,44 @@ def create_recipe(request):
         category = request.POST['category']
         recipe_image = request.FILES['recipe_image']
         user = get_object_or_404(User, pk=request.user.id)
-        recipe = Recipe.objects.create(user=user, recipe_name=recipe_name, ingredients=ingredients, method_of_preparation=method_preparation, time_of_preparation=time_of_preparation, serves=server, category=category, recipe_image=recipe_image)
+        recipe = Recipe.objects.create(user=user, recipe_name=recipe_name,
+                                       ingredients=ingredients,
+                                       method_preparation=method_preparation,
+                                       time_of_preparation=time_of_preparation,
+                                       serves=server,
+                                       category=category,
+                                       recipe_image=recipe_image)
         recipe.save
         return redirect('dashboard')
     else:
         return render(request, 'recipe/create_recipe.html')
 
 
+@login_required
 def delete_recipe(request, recipe_id):
     """ Way to delete a recipe """
-    recipe = get_object_or_404(Recipe, pk=recipe_id)
-    recipe.delete()
-    return redirect('dashboard')
+
+    if request.user.is_authenticated:
+        recipe = get_object_or_404(Recipe, pk=recipe_id)
+
+        if request.method == 'POST':
+            recipe.delete()
+            messages.success(request,
+                             ('Success! Your recipe has been deleted.'))
+            return redirect('dashboard')
+
+        context = {
+            'recipe': recipe,
+            }
+
+        return render(request, 'recipe/delete_recipe.html', context)
+    else:
+        messages.success(resquest, (
+            'You must be signed in to access this page'))
+        return redirect('/final_user/login')
 
 
+@login_required(login_url='/final_user/login')
 def edit_recipe(request, recipe_id):
     """ Button to edit a recipe """
     recipe = get_object_or_404(Recipe, pk=recipe_id)
@@ -62,6 +88,7 @@ def edit_recipe(request, recipe_id):
     return render(request, 'recipe/edit_recipe.html', recipe_to_edit)
 
 
+@login_required(login_url='/final_user/login')
 def update_recipe(request):
     """ Form to edit a recipe and save the edition """
     if request.method == 'POST':
